[CL-2225] Reorganize widgets folder
