+++ front/app/containers/Admin/reporting/components/ReportBuilder/Widgets/AnalyticsChartWidget/index.tsx
@@ -4,7 +4,7 @@ import React from 'react';
 import { Box, Input, Title } from '@citizenlab/cl2-component-library';
 import GenderChart from 'containers/Admin/dashboard/users/Charts/GenderChart';
 import AgeChart from 'containers/Admin/dashboard/users/Charts/AgeChart';
-import VisitorsReportCard from './VisitorsReportCard';
+import VisitorsReportCard from '../VisitorsCard';
 import VisitorsTrafficSourcesReportCard from './VisitorsTrafficSourcesReportCard';
 
 // Utils
